update: server should never try to store models
