+++ python/xpctl/backends/backend.py
@@ -387,16 +387,6 @@ def json2log(events, log_file):
 
 
 @exporter
-def get_checkpoint(checkpoint_base, checkpoint_store, config_sha1, hostname):
-    if checkpoint_base:
-        model_loc = store_model(checkpoint_base, config_sha1, checkpoint_store)
-        if model_loc is not None:
-            return "{}:{}".format(hostname, os.path.abspath(model_loc))
-        else:
-            raise RuntimeError("model could not be stored, see previous errors")
-
-
-@exporter
 def get_experiment_label(config_obj, task, **kwargs):
     if kwargs.get('label', None) is not None:
         return kwargs['label']
@@ -414,41 +404,6 @@ def safe_get(_object, key, alt):
     if val is None or str(val) is None:
         return alt
     return val
-
-
-@exporter
-def store_model(checkpoint_base, config_sha1, checkpoint_store, print_fn=print):
-    checkpoint_base = unzip_model(checkpoint_base)
-    mdir, mbase = os.path.split(checkpoint_base)
-    mdir = mdir if mdir else "."
-    if not os.path.exists(mdir):
-        print_fn("no directory found for the model location: [{}], aborting command".format(mdir))
-        return None
-    
-    mfiles = ["{}/{}".format(mdir, x) for x in os.listdir(mdir) if x.startswith(mbase + "-") or
-              x.startswith(mbase + ".")]
-    if not mfiles:
-        print_fn("no model files found with base [{}] at location [{}], aborting command".format(mbase, mdir))
-        return None
-    model_loc_base = "{}/{}".format(checkpoint_store, config_sha1)
-    if not os.path.exists(model_loc_base):
-        os.makedirs(model_loc_base)
-    dirs = [int(x[:-4]) for x in os.listdir(model_loc_base) if x.endswith(".zip") and x[:-4].isdigit()]
-    # we expect dirs in numbers.
-    new_dir = "1" if not dirs else str(max(dirs) + 1)
-    model_loc = "{}/{}".format(model_loc_base, new_dir)
-    os.makedirs(model_loc)
-    for mfile in mfiles:
-        shutil.copy(mfile, model_loc)
-        print_fn("writing model file: [{}] to store: [{}]".format(mfile, model_loc))
-    print_fn("zipping model files")
-    shutil.make_archive(base_name=model_loc,
-                        format='zip',
-                        root_dir=model_loc_base,
-                        base_dir=new_dir)
-    shutil.rmtree(model_loc)
-    print_fn("model files zipped and written")
-    return model_loc + ".zip"
 
 
 @exporter
