+++ python/xpctl/clients/helpers.py
@@ -1,8 +1,11 @@
+import os
+import shutil
 import pandas as pd
 from swagger_client.models import Experiment, ExperimentAggregate, Result, AggregateResult, TaskSummary
 from typing import List
-from baseline.utils import write_config_file, read_config_file
+from baseline.utils import write_config_file, read_config_file, unzip_model
 import json
+import ast
 
 pd.set_option('display.expand_frame_repr', False)
 pd.set_option("display.max_rows", None)
@@ -176,3 +179,47 @@ def to_experiment(task, configf, logf, user, label):
         test_events=test_events,
         config=config
     )
+
+
+def store_model(checkpoint_base, config_sha1, checkpoint_store, print_fn=print):
+    checkpoint_base = unzip_model(checkpoint_base)
+    mdir, mbase = os.path.split(checkpoint_base)
+    mdir = mdir if mdir else "."
+    if not os.path.exists(mdir):
+        print_fn("no directory found for the model location: [{}], aborting command".format(mdir))
+        return None
+    
+    mfiles = ["{}/{}".format(mdir, x) for x in os.listdir(mdir) if x.startswith(mbase + "-") or
+              x.startswith(mbase + ".")]
+    if not mfiles:
+        print_fn("no model files found with base [{}] at location [{}], aborting command".format(mbase, mdir))
+        return None
+    model_loc_base = "{}/{}".format(checkpoint_store, config_sha1)
+    if not os.path.exists(model_loc_base):
+        os.makedirs(model_loc_base)
+    dirs = [int(x[:-4]) for x in os.listdir(model_loc_base) if x.endswith(".zip") and x[:-4].isdigit()]
+    # we expect dirs in numbers.
+    new_dir = "1" if not dirs else str(max(dirs) + 1)
+    model_loc = "{}/{}".format(model_loc_base, new_dir)
+    os.makedirs(model_loc)
+    for mfile in mfiles:
+        shutil.copy(mfile, model_loc)
+        print_fn("writing model file: [{}] to store: [{}]".format(mfile, model_loc))
+    print_fn("zipping model files")
+    shutil.make_archive(base_name=model_loc,
+                        format='zip',
+                        root_dir=model_loc_base,
+                        base_dir=new_dir)
+    shutil.rmtree(model_loc)
+    print_fn("model files zipped and written")
+    return model_loc + ".zip"
+
+
+def read_config_stream(config):
+    try:
+        return json.loads(config)
+    except json.decoder.JSONDecodeError:
+        try:
+            return ast.literal_eval(config)
+        except ValueError:
+            return None
