+++ python/xpctl/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.5.2"
+__version__ = "1.0.0"
